# elearning/e-smart-learn/app/docs/swagger.yaml
definitions:
...
  dto.AuthorResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_following:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.BlogResponse:
    properties:
      author:
        $ref: '#/definitions/dto.AuthorResponse'
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_read:
        type: boolean
      published_at:
        type: string
      scheduled_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
...
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateBlogRequest:
    properties:
      category_id:
        type: string
      content:
        type: string
      image_url:
        type: string
      scheduled_at:
        type: string
      status:
        enum:
        - draft
        - published
        - scheduled
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - category_id
    - content
    - image_url
    - status
    - tags
    - title
    type: object
...
  dto.UpdateBlogRequest:
    properties:
      category_id:
        type: string
      content:
        type: string
      image_url:
        type: string
      scheduled_at:
        type: string
      status:
        enum:
        - draft
        - published
        - scheduled
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - category_id
    - content
    - image_url
    - status
    - tags
    - title
    type: object
...
info:
  contact: {}
paths:
  /api/v1/admin/blogs:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of blogs with optional filters (admin only)
      parameters:
      - default: 10
        description: limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      - default: created_at
        description: sortBy
        enum:
        - created_at
        - title
        - view_total
        in: query
        name: sortBy
        type: string
      - default: desc
        description: sortOrder
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: category ID
        in: query
        name: category_id
        type: string
      - description: blog status filter
        enum:
        - draft
        - published
        - scheduled
        in: query
        name: status
        type: string
      - description: search keyword (title)
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Get list of blogs with filters (admin)
      tags:
      - blogs
  /api/v1/admin/blogs/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single blog post by its ID (admin only)
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Get a blog post by ID (admin)
      tags:
      - blogs
  /api/v1/admin/blogs/statistics:
    get:
      consumes:
      - application/json
      description: Get blog KPI statistics for current user (total articles, total
        views, average engagement, counts by status)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Get blog dashboard statistics (admin)
      tags:
      - blogs
...
  /api/v1/blogs:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of published blogs with optional filters
      parameters:
      - default: 10
        description: limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: offset
        in: query
        name: offset
        type: integer
      - default: created_at
        description: sortBy
        enum:
        - created_at
        - title
        - view_total
        in: query
        name: sortBy
        type: string
      - default: desc
        description: sortOrder
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: category ID
        in: query
        name: category_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
      summary: Get list of published blogs
      tags:
      - blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post and return created resource
      parameters:
      - description: Create blog payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
      summary: Create a blog post
      tags:
      - blogs
  /api/v1/blogs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post identified by id
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      security:
      - BearerAuth: []
      summary: Delete a blog post
      tags:
      - blogs
    put:
      consumes:
      - application/json
      description: Update a blog post identified by id
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Update blog payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      security:
      - BearerAuth: []
      summary: Update a blog post
      tags:
      - blogs
  /api/v1/blogs/{slug}:
    get:
      consumes:
      - application/json
      description: Returns a blog post identified by its slug, including author info,
        view count and timestamps.
      parameters:
      - description: Blog Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Retrieve a blog post by slug
      tags:
      - blogs
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"